Implement discriminated `Result`
This re-architects the way that `Result` is defined so that the `isOk` and `isErr` properties (yeah, we're back to that now) are capable of narrowing a `Result<T, E>` into either an `OkImpl<T>` or an `ErrImpl<E>`. This solves some type-safety problems that I ran into with the `any` and `all` implementations for `Result` that were previously authored.

Because `Result` is now a union type, rather than an abstract base class, there is no good way to hang the `any` and `all` functions off of it directly; they are now just exposed as plain old functions that can be imported, which works just the same as before, but looses the nice symmetry with the similar `Promise` APIs that the library previously had.