Remove distinct union type and method
This removes the union type and interface for each type and uses just an interface.

The `is___` methods no longer can discriminate between types fully; it only ensures that the instance is a single particular type. This simplifies the architecture of the library and makes documentation easier at the expense of not being able to be as exhaustive about the types in conditionals. For example, an `if/else` condition that checks for `isSome` no longer knows that the `else` condition contains a `None`.

Related to #2